added support for removing unavailable check categories
