+++ src/js/actions/ProjectDetailsActions.js
@@ -5,7 +5,7 @@ import fs from 'fs-extra';
 import ospath from 'ospath';
 // actions
 import * as AlertModalActions from "./AlertModalActions";
-import {getTranslate, getUsername, getProjectSaveLocation, getProjectBookId, getToolCategories} from "../selectors";
+import {getTranslate, getUsername, getProjectSaveLocation, getProjectBookId, getToolCategories, getCurrentProjectToolsSelectedGL} from "../selectors";
 import {cancelProjectValidationStepper} from "./ProjectImportStepperActions";
 // helpers
 import * as bibleHelpers from '../helpers/bibleHelpers';
@@ -28,8 +28,11 @@ const PROJECTS_PATH = path.join(ospath.home(), 'translationCore', 'projects');
  * i.e. ~/translationCore/projects/en_tit_reg
  */
 export const loadCurrentCheckCategories = (toolName, bookName, projectSaveLocation) => {
-  return dispatch => {
-    const selectedCategories = ProjectDetailsHelpers.getCategoriesForProjectFromFS(toolName, bookName, projectSaveLocation);
+  return (dispatch, getState) => {
+    const currentProjectToolsSelectedGL = getCurrentProjectToolsSelectedGL(getState());
+    const availableCheckCategories = ProjectDetailsHelpers.getAvailableCheckCategories(currentProjectToolsSelectedGL);
+    let selectedCategories = ProjectDetailsHelpers.getCachedCategoriesFromProject(toolName, bookName, projectSaveLocation);
+    selectedCategories = selectedCategories.filter((category) => availableCheckCategories[toolName] && availableCheckCategories[toolName].includes(category));
     dispatch(setCategories(selectedCategories, toolName));
   };
 };
